requirements.txt
afwf==0.0.3
atomicwrites==1.4.0
attrs==21.4.0
attrs-mate==1.0.2
autopep8==1.6.0
boto3==1.21.39
botocore==1.24.39
diskcache==5.4.0
jmespath==1.0.0
pathlib-mate==1.0.3
pycodestyle==2.8.0
python-dateutil==2.8.2
s3pathlib==1.0.6
s3transfer==0.5.2
six==1.16.0
smart-open==5.2.1
sqlitedict==1.7.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
toml==0.10.2
urllib3==1.26.9
